fix bug that tables are not created on plugin activation #108
Despite of activation of the freecycle plugin, custmize tables were not created.
It was because of the incorrect definition of event hook of the plugin activation.
On register_activation_hook function, the full path of main plugin php file should be passed as the first argument. But the path of freecycle_meta_table.php was passed instead of freecycle.php. This should have been changed when the main plugin php was changed to freecyle.php from freecycle_meta_table.php.
To fix this, delete register_activation_hook function in freecycle_meta_table.php and call fmt_activation function directly, which was called from register_activation_hook before.
Besides, fix some sql errors which occurs when the plugin is activated on the environment already having freecycle plugin.