Fix bug #94
This bug occured only when WP_DEBUG_DISPLAY was true in wp-config.php.
It was because it was not checked if 'keyword' value was set in $_GET.
To fix this, isset method was implemented.
Basically, isset method should be used anytime any values specified by name are taken from allay.